[modules] nav_catapult: set delay in seconds untested, but should close #188

+++ sw/airborne/modules/nav/nav_catapult.h
@@ -32,9 +32,9 @@
 #include "std.h"
 #include "paparazzi.h"
 
-extern int nav_catapult_motor_delay;
+extern float nav_catapult_motor_delay;
 extern float nav_catapult_acceleration_threshold;
-extern int nav_catapult_heading_delay;
+extern float nav_catapult_heading_delay;
 extern float nav_catapult_initial_pitch;
 extern float nav_catapult_initial_throttle;
